% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{get_camels_path}
\alias{get_camels_path}
\title{Get CAMELS-PE dataset path}
\usage{
get_camels_path()
}
\value{
Character string with the CAMELS-PE root directory.
}
\description{
Returns the CAMELS-PE root directory previously defined with
\code{set_camels_path()}.
}
\examples{
path <- system.file(
  "extdata",
  "sample_camels_pe",
  package = "RCamelsPE"
)

# Define CAMELS-PE root directory
set_camels_path(path)

# Retrieve the stored path
get_camels_path()
}
