% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_catchments}
\alias{plot_catchments}
\title{Plot CAMELS-PE Catchments}
\usage{
plot_catchments(catchments, gauges = NULL, fill = NULL, ...)
}
\arguments{
\item{catchments}{An \code{sf} object with catchment polygons, typically
obtained with \code{read_geospatial("catchments")}.}

\item{gauges}{Optional \code{sf} object with gauge point locations, typically
obtained with \code{read_geospatial("gauges")}.}

\item{fill}{Optional character string. Name of a catchment column used to
fill polygons. If NULL, catchments are plotted with a constant fill.}

\item{...}{Additional arguments passed to \code{ggplot2::geom_sf()} for
catchments.}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a simple map of CAMELS-PE catchments and, optionally, gauge
locations. This function is useful for quick inspection of the spatial
distribution of catchments and stations.
}
\examples{
\dontrun{
catchments <- read_geospatial("catchments")
gauges <- read_geospatial("gauges")

plot_catchments(catchments)
plot_catchments(catchments, gauges)
}

}
