% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path.R
\name{set_camels_path}
\alias{set_camels_path}
\title{Set CAMELS-PE dataset path}
\usage{
set_camels_path(path)
}
\arguments{
\item{path}{Character. Path to the CAMELS-PE root directory.}
}
\value{
Invisibly returns the normalized path.
}
\description{
Set CAMELS-PE dataset path
}
