% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_pe_data}
\alias{download_pe_data}
\title{Download the CAMELS-PE dataset from Zenodo}
\usage{
download_pe_data(
  dest_dir = tools::R_user_dir("rcamelspe", which = "data"),
  unzip = TRUE,
  overwrite = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{dest_dir}{Character string. Destination directory where the dataset should be
saved. Defaults to the persistent user data directory
(\code{tools::R_user_dir("rcamelspe", "data")}).}

\item{unzip}{Logical. Should the downloaded file be unzipped? Default is \code{TRUE}.}

\item{overwrite}{Logical. If destination files already exist, should they be overwritten?
Default is \code{FALSE}.}

\item{quiet}{Logical. Should download progress be suppressed? Default is \code{FALSE}.}
}
\value{
Character string indicating the directory path where the dataset is located
or extracted.
}
\description{
Downloads the CAMELS-PE dataset zip file from the Zenodo repository
(\doi{10.5281/zenodo.20058779}) and extracts it in the destination folder.
}
\examples{
\dontrun{
# Download and unzip to persistent user data folder
download_pe_data()

# Download to a custom folder
download_pe_data(dest_dir = tempdir())
}
}
