% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paths.R
\name{get_camels_pe_path}
\alias{get_camels_pe_path}
\title{Get the path to the CAMELS-PE dataset}
\usage{
get_camels_pe_path()
}
\value{
Character path or \code{NULL} if not found.
}
\description{
Retrieves the path to the CAMELS-PE dataset. It looks up:
\enumerate{
\item The package option \code{rcamelspe.path}.
\item The environment variable \code{CAMELS_PE_PATH}.
\item Default search paths in the current working directory (\code{raw-data/CAMELS-PE}, \code{data-raw/camels-pe}, etc.).
}
}
\examples{
get_camels_pe_path()
}
